chore(siwe): apply review feedback from fbac
- Drop unused retryCount param on authenticateWithSIWE
  (protocol, impl, 6 test conformers, Mock, BackendAuthProbe,
  SessionStateMachine call site). The fresh-nonce retry budget is
  internal to authenticateWithSIWE; callers don't drive it.
- Per-env SIWE values in config.{env}.json:
    local / dev / pr -> dev.convos.org
    prod             -> convos.org
- AppEnvironment .tests case now uses a `tests.invalid` placeholder
  with chainId 0 so it can't be confused with a real env. Tests
  that exercise SIWE construct their own SIWEMessage inputs and
  never read this accessor.
- Split cookie parsing into its own file
  (ConvosAPIClient+SIWECookieParsing.swift). The SIWE flow file
  keeps the request/response orchestration; cookie semantics live
  next to their own tests.
