+++ Convos/Config/config.local.json
@@ -10,7 +10,7 @@
     "app-local.convos.org"
   ],
   "appUrlScheme": "convos-local",
-  "siweDomain": "convos.app",
-  "siweUri": "https://convos.app",
+  "siweDomain": "dev.convos.org",
+  "siweUri": "https://dev.convos.org",
   "siweChainId": 1
 }
