+++ ConvosCore/Tests/ConvosCoreTests/Integration/PushTopicSubscriptionDeniedDMTests.swift
@@ -154,7 +154,7 @@ private final class ThrowawayPushAPIClient: ConvosAPIClientProtocol, @unchecked 
     func registerDevice(deviceId: String, pushToken: String?) async throws {}
 
     func authenticate(appCheckToken: String, retryCount: Int) async throws -> String { "token" }
-    func authenticateWithSIWE(appCheckToken: String, signing: BackendAuthSigningContext, retryCount: Int) async throws -> String { "siwe-token" }
+    func authenticateWithSIWE(appCheckToken: String, signing: BackendAuthSigningContext) async throws -> String { "siwe-token" }
     func updateSIWESigningContext(_ context: BackendAuthSigningContext?) {}
     func accountAuthCheck(jwt: String?) async throws -> ConvosAPI.AuthCheckResponse { .init(success: jwt != nil) }
