+++ ConvosCore/Tests/ConvosCoreTests/SyncingManagerTests.swift
@@ -346,7 +346,7 @@ final class TestableMockAPIClient: ConvosAPIClientProtocol, @unchecked Sendable 
         "mock-jwt-token"
     }
 
-    func authenticateWithSIWE(appCheckToken: String, signing: BackendAuthSigningContext, retryCount: Int) async throws -> String {
+    func authenticateWithSIWE(appCheckToken: String, signing: BackendAuthSigningContext) async throws -> String {
         "mock-siwe-jwt-token"
     }
